requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-sso==0.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
load-dotenv==0.1.0
oauthlib==3.2.2
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
